Add source files for templates and styling
I used Nunjucks as the main template language, and Sass with Gorko for
styling and design structure, following the principles of CUBE CSS. This
is only a single page site, so I added the CSS as an internal stylesheet
to minimize requests.

During the build process, Sass compiles the files in the `scss`
directory into a `css` folder in the `_includes` directory. Then, 11ty
processes the template files, which includes the `global.css` file in
the <head> of the page.

Fun fact: this is the first project where I tried using Gorko. I don't
think I was able to maximize its features, but I now realize the power
of using such tool and thinking in terms of design tokens.